clean up helpers in internal/cmd
